Splitting services 1.Fixes for foreign keys resolved in rpc calls (it neeeds even more fixes) 2.Now we use TThreadedPoolServer which is non-blocking 3.Acrhitecture diagram updated with current ER

+++ DBServices/RobotsService/src/main/java/com/qreal/robots/server/RobotDbServiceHandler.java
@@ -28,19 +28,32 @@ public class RobotDbServiceHandler implements RobotDbService.Iface {
     @Override
     public TRobot findById(long id) throws TException {
         RobotDao robotDao = (RobotDao) context.getBean("robotDao");
-        return robotDao.findById(id).toTRobot();
+        RobotSerial robot = robotDao.findById(id);
+        if (robot != null) {
+            return robot.toTRobot();
+        }
+        return null;
     }
 
     @Override
     public void deleteRobot(long id) throws TException {
         RobotDao robotDao = (RobotDao) context.getBean("robotDao");
         RobotSerial robot = robotDao.findById(id);
-        robotDao.delete(robot);
+        if (robot != null) {
+            robotDao.delete(robot);
+        }
     }
 
     @Override
     public boolean isRobotExists(long id) throws TException {
         RobotDao robotDao = (RobotDao) context.getBean("robotDao");
         return robotDao.isRobotExists(id);
+    }
+
+    @Override
+    public void updateRobot(TRobot tRobot) throws TException {
+        RobotDao robotDao = (RobotDao) context.getBean("robotDao");
+        RobotSerial robotSerial = new RobotSerial(tRobot);
+        robotDao.updateRobot(robotSerial);
     }
 }
